+++ iex-parser/iex_cppparser/__init__.py
@@ -1,21 +1,21 @@
 import os
-from datetime import timedelta, datetime
+from datetime import datetime, timedelta
 from .download import download_hist_file
 import glob
 import subprocess
 import argparse
-from datetime import datetime
 
 # Path to the directory this package is installed in. Used for base path for running C++ binary files
 dir_path = os.path.dirname(os.path.realpath(__file__))
 
+
 def valid_date(s: str) -> str:
     """
     This function checks if a given string represents a valid date in the format YYYY-MM-DD.
-    
+
     Parameters:
         s (str): The string to be checked.
-    
+
     Returns:
         s (str): The input string if it represents a valid date, otherwise raises an error.
     """
@@ -29,7 +29,7 @@ def valid_date(s: str) -> str:
 def parse_file(file_path: str, parsed_folder: str, symbol: str, split: bool = False):
     """
     This function parses a file using the IEX parser and redirects the output to a specified folder.
-    
+
     Parameters:
         file_path (str): The path to the file to be parsed.
 
@@ -38,7 +38,7 @@ def parse_file(file_path: str, parsed_folder: str, symbol: str, split: bool = Fa
         symbol (str): Path to a txt file with symbols to parse. Must have one symbol per line. If "ALL", all symbols are parsed.
 
         split (bool): Whether to split the output files. One file per letter of the alphabet is generated. Default is False. Using split=True will slow down the parsing process as it will not be multi-threaded. USe this only if you have memory constraints.
-        
+
     Returns:
         None
 
@@ -46,30 +46,38 @@ def parse_file(file_path: str, parsed_folder: str, symbol: str, split: bool = Fa
         Two files are generated:
 
         - The file ending in `_trd.csv` contains the trades data.
-        
+
         - The file ending in `_prl.csv` contains the price level updates.
 
     """
-    if split==True:
+    if split:
         # Use compiled C++ binary to parse and split output files
-        IEX_PARSER = os.path.join(dir_path, 'bin/iex_parser_split.out')
+        IEX_PARSER = os.path.join(dir_path, "bin/iex_parser_split.out")
     else:
         if symbol == "ALL":
             # Use compiled C++ binary to parse ALL symbols
-            IEX_PARSER = os.path.join(dir_path, 'bin/iex_parser_all_threaded.out')
+            IEX_PARSER = os.path.join(dir_path, "bin/iex_parser_all_threaded.out")
         else:
             # Use compiled C++ binary to parse selected symbols
-            IEX_PARSER =  os.path.join(dir_path, 'bin/iex_parser_threaded.out')
-    
+            IEX_PARSER = os.path.join(dir_path, "bin/iex_parser_threaded.out")
+
     parsed_prefix = os.path.join(parsed_folder, os.path.basename(file_path).replace(".pcap.gz", ""))
-        
-    command2 =f"gunzip -d -c {file_path} | tcpdump -r - -w - -s 0 |  {IEX_PARSER} /dev/stdin {parsed_prefix} {symbol}"
+
+    command2 = f"gunzip -d -c {file_path} | tcpdump -r - -w - -s 0 |  {IEX_PARSER} /dev/stdin {parsed_prefix} {symbol}"
     subprocess.run(command2, shell=True)
 
-def parse_date(date_str: str, download_dir: str, parsed_folder: str, symbol: str, download: bool = True, split: bool = False):
+
+def parse_date(
+    date_str: str,
+    download_dir: str,
+    parsed_folder: str,
+    symbol: str,
+    download: bool = True,
+    split: bool = False,
+):
     """
     This function (can) download and parse the IEXTP1 DEEP1.0 pcap files for a given date.
-    
+
     Parameters:
         date_str (str): The date string to be parsed. Format YYYY-MM-DD
 
@@ -82,7 +90,7 @@ def parse_date(date_str: str, download_dir: str, parsed_folder: str, symbol: str
         download (bool): Whether to download the files. Default is True.
 
         split (bool): Whether to split the output files. One file per letter of the anphabet is generated. Default is False. Using split=True will slow down the parsing process as it will not be multi-threaded. USe this only if you have memory constraints.
-        
+
     Returns:
         None
 
@@ -97,7 +105,7 @@ def parse_date(date_str: str, download_dir: str, parsed_folder: str, symbol: str
     if valid_date(date_str) is None:
         return
 
-    date_str_2 = date_str.replace("-","")
+    date_str_2 = date_str.replace("-", "")
     file_pattern = f"data_feeds_{date_str_2}_{date_str_2}_IEXTP1_DEEP1.0.pcap.gz"
 
     if download:
@@ -106,13 +114,21 @@ def parse_date(date_str: str, download_dir: str, parsed_folder: str, symbol: str
     matching_files = glob.glob(f"{download_dir}/{file_pattern}")
 
     for file_path in matching_files:
-        parse_file(file_path, parsed_folder, symbol,split=split)
+        parse_file(file_path, parsed_folder, symbol, split=split)
 
 
-def parse_dates(start_date: str, end_date: str, download_dir: str, parsed_folder: str, symbol: str, download: bool = True, split: bool = False):
+def parse_dates(
+    start_date: str,
+    end_date: str,
+    download_dir: str,
+    parsed_folder: str,
+    symbol: str,
+    download: bool = True,
+    split: bool = False,
+):
     """
     This function parses a range of dates and (downloads and) parses the corresponding IEXTP1 DEEP1.0 pcap files.
-    
+
     Parameters:
         start_date (str): The start date string in the format YYYY-MM-DD.
 
@@ -125,15 +141,15 @@ def parse_dates(start_date: str, end_date: str, download_dir: str, parsed_folder
         symbol (str): Path to a txt file with symbols to parse. Must have one symbol per line. If "ALL", all symbols are parsed.
 
         download (bool): Whether to download the files. Default is False.
-        
+
         split (bool): Whether to split the output files. One file per letter of the anphabet is generated. Default is False. Using split=True will slow down the parsing process as it will not be multi-threaded. USe this only if you have memory constraints.
-    
+
     Returns:
         None
 
     Output:
         For each date, two files are generated:
-        
+
         - The file ending in `_trd.csv` contains the trades data.
 
         - The file ending in `_prl.csv` contains the price level updates.
@@ -147,10 +163,9 @@ def parse_dates(start_date: str, end_date: str, download_dir: str, parsed_folder
 
     current_date = start_date
     while current_date <= end_date:
-
         current_date_str = current_date.strftime("%Y-%m-%d")
         try:
-            parse_date(current_date_str, download_dir, parsed_folder, symbol,download,split=split)
+            parse_date(current_date_str, download_dir, parsed_folder, symbol, download, split=split)
         except Exception as e:
             print(f"Error parsing date {current_date_str}: {e}")