+++ iex-parser/iex_cppparser/which_bash.py
@@ -1,6 +1,6 @@
-import shutil
 import subprocess
 from pathlib import Path
+
 
 def find_bash_path():
     try:
@@ -20,20 +20,20 @@ def write_bash_path_to_env_file(env_file_path, bash_path):
         env_file = Path(env_file_path)
         if env_file.exists():
             # Read existing content
-            with open(env_file, 'r') as f:
+            with open(env_file, "r") as f:
                 existing_content = f.read()
                 if f"BASH_PATH={bash_path}" in existing_content:
                     print(f"Bash path already exists in {env_file_path}.")
                     return
 
         # Append the Bash path to the environment file
-        with open(env_file, 'a') as f:
+        with open(env_file, "a") as f:
             f.write(f"BASH_PATH={bash_path}\n")
         print(f"Bash path written to {env_file_path}")
     except Exception as e:
         print(f"Error writing to {env_file_path}: {e}")
 
+
 if __name__ == "__main__":
     write_bash_path_to_env_file("/vagrant/.env", find_bash_path())
     print(f"The path to the Bash interpreter is: {find_bash_path()}")
-